IMPALA-12264: Add limit on number of HS2 sessions per user.
Add a new flag --max_hs2_sessions_per_user which sets a limit on the
number of Hiveserver2 sessions that can be concurrently opened by a
user on a coordinator. By default this value is zero, which disables
the new feature. An attempt to open more sessions than the new flag
value results in an error. This is implemented by maintaining a map of
users to a count of HS2 sessions.

Expose the per-user HS2 session counts on the /sessions page of the
webui.

Add a new test case in test_session_expiration.py.

Change-Id: Idd28edc352102d89774f6ece5376e7c79ae41aa8