IMPALA-12943: Document Admission Control User Quotas.
Document the feature introduced in IMPALA-12345. Add a few more tests to
the QuotaExamples test which demonstrate the examples used in the
docs.

Clarify in docs and code the behavior when a user is a member of more
than one group for which there are rules. In this case the least
restrictive rule applies.

Also document the '--max_hs2_sessions_per_user' flag introduced in
IMPALA-12264.

Change-Id: I82e044adb072a463a1e4f74da71c8d7d48292970